tests: Follow Git index helper module
The utility tests have separate modules for command execution and
repository helpers, while temporary index plumbing remains in the command
utility tests.

The source index helpers now live in utils.git_index. Tests that keep those
cases in test_git.py obscure the new module boundary and import moved names
from the old command wrapper.

This commit addresses that mismatch by moving temporary index plumbing
coverage into tests/utils/test_git_index.py and removing stale index
imports from tests/utils/test_git.py.

The final commit will add architecture coverage for the Git index helper
boundary.
